move course components into components/courses
