--- src/components/courseDetail.js
@@ -1,385 +1,0 @@
-import React from 'react';
-import classNames from 'classnames';
-import PropTypes from 'prop-types';
-import { withStyles } from 'material-ui/styles';
-import Table, {
-  TableBody,
-  TableCell,
-  TableFooter,
-  TableHead,
-  TablePagination,
-  TableRow,
-  TableSortLabel,
-} from 'material-ui/Table';
-import { compose } from 'redux'
-import { connect } from 'react-redux'
-import { firebaseConnect, isLoaded, isEmpty } from 'react-redux-firebase'
-import Toolbar from 'material-ui/Toolbar';
-import Typography from 'material-ui/Typography';
-import Paper from 'material-ui/Paper';
-import Checkbox from 'material-ui/Checkbox';
-import IconButton from 'material-ui/IconButton';
-import Tooltip from 'material-ui/Tooltip';
-import DeleteIcon from 'material-ui-icons/Delete';
-import FilterListIcon from 'material-ui-icons/FilterList';
-import { lighten } from 'material-ui/styles/colorManipulator';
-import Snackbar from 'material-ui/Snackbar';
-import Tabs, { Tab } from 'material-ui/Tabs';
-import Button from 'material-ui/Button';
-import CreateAssignment from './createAssignment'
-import AppFrame from '../AppFrame'
-
-let counter = 0;
-function createData(name, calories, fat, carbs, protein) {
-  counter += 1;
-  return { id: counter, name, calories, fat, carbs, protein };
-}
-
-const columnData = [
-  { id: 'name', numeric: false, disablePadding: true, label: 'Name' },
-  { id: 'calories', numeric: true, disablePadding: false, label: 'Description' }
-];
-
-class EnhancedTableHead extends React.Component {
-  createSortHandler = property => event => {
-    this.props.onRequestSort(event, property);
-  };
-
-  render() {
-    const { onSelectAllClick, order, orderBy, numSelected, rowCount } = this.props;
-
-    return (
-      <TableHead>
-        <TableRow>
-          <TableCell padding="checkbox">
-            <Checkbox
-              indeterminate={numSelected > 0 && numSelected < rowCount}
-              checked={numSelected === rowCount}
-              onChange={onSelectAllClick}
-            />
-          </TableCell>
-          {columnData.map(column => {
-            return (
-              <TableCell
-                key={column.id}
-                numeric={column.numeric}
-                padding={column.disablePadding ? 'none' : 'default'}
-                sortDirection={orderBy === column.id ? order : false}
-              >
-                <Tooltip
-                  title="Sort"
-                  placement={column.numeric ? 'bottom-end' : 'bottom-start'}
-                  enterDelay={300}
-                >
-                  <TableSortLabel
-                    active={orderBy === column.id}
-                    direction={order}
-                    onClick={this.createSortHandler(column.id)}
-                  >
-                    {column.label}
-                  </TableSortLabel>
-                </Tooltip>
-              </TableCell>
-            );
-          }, this)}
-        </TableRow>
-      </TableHead>
-    );
-  }
-}
-
-EnhancedTableHead.propTypes = {
-  numSelected: PropTypes.number.isRequired,
-  onRequestSort: PropTypes.func.isRequired,
-  onSelectAllClick: PropTypes.func.isRequired,
-  order: PropTypes.string.isRequired,
-  orderBy: PropTypes.string.isRequired,
-  rowCount: PropTypes.number.isRequired,
-};
-
-const toolbarStyles = theme => ({
-  root: {
-    paddingRight: theme.spacing.unit,
-  },
-  title: {
-    flex: '0 0 auto',
-  },
-});
-
-let EnhancedTableToolbar = props => {
-  const { numSelected, classes } = props;
-
-  return (
-    <Toolbar
-      className={classNames(classes.root, {
-        [classes.highlight]: numSelected > 0,
-      })}
-    >
-      <div className={classes.title}>
-        {numSelected > 0 ? (
-          <Typography type="subheading">{numSelected} selected</Typography>
-        ) : (
-          <Typography type="title">Course Detail</Typography>
-        )}
-      </div>
-      <div className={classes.spacer} />
-      <div className={classes.actions}>
-        {numSelected > 0 ? (
-          <Tooltip title="Delete">
-            <IconButton aria-label="Delete">
-              <DeleteIcon />
-            </IconButton>
-          </Tooltip>
-        ) : (
-          <Tooltip title="Filter list">
-            <IconButton aria-label="Filter list">
-              <FilterListIcon />
-            </IconButton>
-          </Tooltip>
-        )}
-      </div>
-    </Toolbar>
-  );
-};
-
-EnhancedTableToolbar.propTypes = {
-  classes: PropTypes.object.isRequired,
-  numSelected: PropTypes.number.isRequired,
-};
-
-EnhancedTableToolbar = withStyles(toolbarStyles)(EnhancedTableToolbar);
-
-const styles = theme => ({
-  root: {
-    width: '100%',
-    marginTop: theme.spacing.unit * 3,
-  },
-  table: {
-    minWidth: 800,
-  },
-  tableWrapper: {
-    overflowX: 'auto',
-  },
-});
-
-class CourseDetail extends React.Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {
-      order: 'asc',
-      orderBy: 'calories',
-      selected: [],
-      data: null,
-      page: 0,
-      rowsPerPage: 5,
-      isAsgmtActive:false,
-      open: false,
-      vertical: 'top',
-      horizontal: 'right',
-      message:null,
-      value: 0
-    };
-  }
-
-  handleRequestSort = (event, property) => {
-    const orderBy = property;
-    let order = 'desc';
-
-    if (this.state.orderBy === property && this.state.order === 'desc') {
-      order = 'asc';
-    }
-
-    const data =
-      order === 'desc'
-        ?this.props.assignment.sort((a, b) => (b[orderBy] < a[orderBy] ? -1 : 1))
-        : this.props.assignment.sort((a, b) => (a[orderBy] < b[orderBy] ? -1 : 1));
-
-    this.setState({ data, order, orderBy });
-  };
-
-  handleSelectAllClick = (event, checked) => {
-    if (checked) {
-      this.setState({ selected: this.props.assignment.map(n => n.id) });
-      return;
-    }
-    this.setState({ selected: [] });
-  };
-
-  handleClick = (event, id) => {
-    const { selected } = this.state;
-    const selectedIndex = selected.indexOf(id);
-    let newSelected = [];
-
-    if (selectedIndex === -1) {
-      newSelected = newSelected.concat(selected, id);
-    } else if (selectedIndex === 0) {
-      newSelected = newSelected.concat(selected.slice(1));
-    } else if (selectedIndex === selected.length - 1) {
-      newSelected = newSelected.concat(selected.slice(0, -1));
-    } else if (selectedIndex > 0) {
-      newSelected = newSelected.concat(
-        selected.slice(0, selectedIndex),
-        selected.slice(selectedIndex + 1),
-      );
-    }
-
-    this.setState({ selected: newSelected });
-  };
-  closeAssignment=()=>{
-    this.setState({isAsgmtActive:false})
-  }
-  handleNotification = (msg) =>{
-    this.setState({ open: true,message:msg });
-  };
-  handleClose = () => {
-    this.setState({ open: false });
-  };
-  submitAssignment=(e)=>{
-   let allAssignment={
-      name:e.name,
-      desc:e.desc
-    }
-    
-    this.props.firebase.push('assignment', allAssignment).then( data => {
-      // wait for db to send response\
-
-      this.handleNotification('Data Save Successfully');
-      this.closeAssignment();
-    }) ;
-  }
-
-  createAssignment=()=>{
-    this.setState({isAsgmtActive:true})
-  }
-
-  handleChange = (event, value) => {
-    this.setState({ value });
-  };
-
-  handleChangePage = (event, page) => {
-    this.setState({ page });
-  };
-
-  handleChangeRowsPerPage = event => {
-    this.setState({ rowsPerPage: event.target.value });
-  };
-
-  isSelected = id => this.state.selected.indexOf(id) !== -1;
-
-  render() {
-    const { classes, assignment, auth } = this.props;
-    const { order, orderBy, selected, rowsPerPage, page, vertical, horizontal, open, message  } = this.state;
-    const emptyRows =  assignment ? rowsPerPage - Math.min(rowsPerPage, assignment.length - page * rowsPerPage) :'';
-
-    return (
-      <div>
-      <AppFrame>
-
-        {auth.emailVerified && assignment ?
-      <Paper className={classes.root}>
-      <Tabs
-        value={this.state.value}
-        onChange={this.handleChange}
-        indicatorColor="primary"
-        textColor="primary"
-        centered
-      >
-        <Tab label="ASSIGNMENTS" />
-        <Tab label="EDIT" />
-        <Tab label="INSTRUCTOR VIEW" />
-      </Tabs>
-        <EnhancedTableToolbar numSelected={selected.length} />
-        {this.state.value==0 && <div className={classes.tableWrapper}>
-          <Table className={classes.table}>
-            <EnhancedTableHead
-              numSelected={selected.length}
-              order={order}
-              orderBy={orderBy}
-              onSelectAllClick={this.handleSelectAllClick}
-              onRequestSort={this.handleRequestSort}
-              rowCount={assignment.length}
-            />
-            <TableBody>
-              {assignment.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((n,id) => {
-                const isSelected = this.isSelected(id);
-                return (
-                  <TableRow
-                    hover
-                    onClick={event => this.handleClick(event, id)}
-                    role="checkbox"
-                    aria-checked={isSelected}
-                    tabIndex={-1}
-                    key={id}
-                    selected={isSelected}
-                  >
-                    <TableCell padding="checkbox">
-                      <Checkbox checked={isSelected} />
-                    </TableCell>
-                    <TableCell padding="none">{n.value.name}</TableCell>
-                    <TableCell numeric>{n.value.desc}</TableCell>
-                  </TableRow>
-                );
-              })}
-              {emptyRows > 0 && (
-                <TableRow style={{ height: 49 * emptyRows }}>
-                  <TableCell colSpan={6} />
-                </TableRow>
-              )}
-            </TableBody>
-            <TableFooter>
-              <TableRow>
-                <TablePagination
-                  colSpan={6}
-                  count={assignment.length}
-                  rowsPerPage={rowsPerPage}
-                  page={page}
-                  backIconButtonProps={{
-                    'aria-label': 'Previous Page',
-                  }}
-                  nextIconButtonProps={{
-                    'aria-label': 'Next Page',
-                  }}
-                  onChangePage={this.handleChangePage}
-                  onChangeRowsPerPage={this.handleChangeRowsPerPage}
-                />
-              </TableRow>
-            </TableFooter>
-          </Table>
-        </div>}
-
-        {this.state.value==1 && <Button raised color="primary" onClick={this.createAssignment}>
-          Create a assignment
-      </Button>}
-
-      </Paper> : '' } 
-      { this.state.isAsgmtActive && <CreateAssignment 
-      handleClose={this.closeAssignment}
-      handleSubmit={this.submitAssignment}
-      /> }
-      </AppFrame>
-     
-      <Snackbar
-      anchorOrigin={{ vertical, horizontal }}
-      open={open}
-      onClose={this.handleClose}
-      SnackbarContentProps={{
-        'aria-describedby': 'message-id',
-      }}
-      message={<span id="message-id">{message}</span>}
-    />
-     </div>
-    );
-  }
-}
-
-const CoursesWithFirebase = compose(
-  firebaseConnect(['assignment']),
-  connect(({ firebase }) => ({ auth: firebase.auth, assignment: firebase.ordered.assignment }))
-)(CourseDetail)
-
-CourseDetail.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(CoursesWithFirebase);
