+++ src/components/courses/index.js
@@ -1,0 +1,5 @@
+export * from './courseDetail';
+export * from './createCourse';
+export * from './joinedCourses';
+export * from './myCourses';
+export * from './publicCourses';
