+++ src/components/courses/joinedCourses.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const JoinedCourses = () => (
+    <h2>Joined Course</h2>
+)
