+++ src/components/courses/publicCourses.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export const PublicCourses = () => (
+    <h2>Public Course</h2>
+)
+
