--- src/components/joinedCourses.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const JoinedCourses = () => (
-    <h2>Joined Course</h2>
-)
-
-export default JoinedCourses;
