--- src/components/publicCourses.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const PublicCourses = () => (
-    <h2>Public Course</h2>
-)
-
-export default PublicCourses;
